requirement.txt
beautifulsoup4
lxml
pyenchant